chore: add custom hook into the location screen
